+++ NEWS.md
@@ -1,3 +1,8 @@
-# naturalistbot (development version)
+# naturalistbot (0.0.1)
+
+* Use hierarchical clustering to try to get different types of observations.
+* Include GPS coordinates in the alt text of the map.
+
+# naturalistbot (0.0.0.9000)
 
 * Initial release.